Created utilities.py to keep utility functions
